refactor: decouple provider requests from views and expose the API
