+++ src/plugins/GoogleDrive/index.js
@@ -1,6 +1,6 @@
 const Plugin = require('../../core/Plugin')
-const Provider = require('../Provider')
-const View = require('../Provider/view')
+const { Provider } = require('../../server')
+const { ProviderView } = require('../../views')
 const { h } = require('preact')
 
 module.exports = class GoogleDrive extends Plugin {
@@ -33,7 +33,7 @@ module.exports = class GoogleDrive extends Plugin {
   }
 
   install () {
-    this.view = new View(this)
+    this.view = new ProviderView(this)
     // Set default state for Google Drive
     this.setPluginState({
       authenticated: false,
