+++ src/server/index.js
@@ -1,0 +1,11 @@
+'use-strict'
+/**
+ * Manages communications with Uppy Server
+ */
+
+const Request = require('./Request')
+const Provider = require('./Provider')
+
+module.exports = {
+  Request, Provider
+}
